Move custom properties to their own file
